+++ app/javascript/vue/tasks/field_occurrences/new/components/HeaderBar.vue
@@ -273,6 +273,13 @@ const hotkeys = ref([
     handler() {
       reset()
     }
+  },
+  {
+    keys: [platformKey(), 'l'],
+    preventDefault: true,
+    handler() {
+      settings.toggleLock()
+    }
   }
 ])
 
@@ -291,6 +298,12 @@ TW.workbench.keyboard.createLegend(
 TW.workbench.keyboard.createLegend(
   `${platformKey()}+r`,
   'Reset all',
+  'New field occurrence'
+)
+
+TW.workbench.keyboard.createLegend(
+  `${platformKey()}+l`,
+  'Lock/Unlock all',
   'New field occurrence'
 )
 </script>